Reorganized test_runner test apps.
